refactor: move clubs api methods to service folder
